feat(ci): switch nix-github-actions for a custom solution
This python script runs nix-eval-jobs, and parses the output into a format suitable for github action matrix consumption. It is similar to nix-github-actions, but it actually evals the derivations (thus avoiding each job in the matrix having to re-evaluate), and even checks if they're cached or not (allowing for skipping jobs instantly).

Thanks to this, I guess it's safe to remove the merge-queue skip thing, as pipelines are fully cached take at most 30s to finish.