chore(site): remove PDF meeting notes
Not all meeting notes can reliable turn into PDF with beamer, which are failing silently. It frequently fails in notes that are too big or that have too deep lists, making it a pain for peoplw writing them.

Let's remove them for now, and later I'll replace them with typst.